The procbar.c the diskusage applet uses is not the same as the procbar.c
The procbar.c the diskusage applet uses is not the same as
the procbar.c the cpumemusage applet uses.

I took the file from there, but I added support
for horizontal or vertical bars.

therfore, ../cpumemusage/procbar.c in Makefile.am was wrong.
Strangly enough, it worked nevertheless. Probably
because further down, procbar.o was specified, and automake
choose'd to take ./procbar.c