added patch by Dave Finton <dfinton@d.umn.edu>
I wrote up a little something for the disk usage applet in GNOME.  I
added a sub-menu to the right-click menu (above About... and Properties)
that lists all the mounted filesystems.  Of course, when you select the
menu item, the applet then displays the file system status for your
selected fs.

It automatically updates when fs's are added or removed.