Warning hunt complete.
Warning hunt complete.

Well, that's the last of them. It builds cleanly on my box now...

There are a few I didn't touch, for example:
struct foo bar{0};

This generates a warning, and there are a few of them, but I didn't want
to step on any toes to fix it :)