Added the ebox to the GSwitchItApplet structure because we need it in
2004-02-29  Dennis Smit  <ds@nerds-incorporated.org>

* gswitchit-applet.h: Added the ebox to the
GSwitchItApplet structure because we need it
in GSwitchItAppletChangeBackground.

* gswitchit-applet.c (GSwitchItAppletPrepareDrawing):
Put the label in the align first before putting the
align in the ebox. Else the panel background doesn't
work, why this is i don't know.

(GSwitchItAppletChangeBackground):
Set the GtkStyle on the ebox as well.

The applet now supports panel background colors.
Pixmap isn't working yet tho.