windowpicker/task-item.c: Improve size, position and gradient of the glow pattern
This patch fixes a problem that causes the glow pattern to not apprear centered,
when the area.y value is greater than 0 (was only reproducible in GtkInspector).

This slightly reduces the maximum size of the glow pattern to avoid it from overflowing
the visible area.

This patch increases the gradient distance from the center to make the glow pattern
more visible.