windowpicker: let TaskList only show windows from same monitor
This patch makes a given tasklist ignore windows that
are not on the same monitor as the tasklist, if there is more than one
task list added to gnome-panel. If a window moves from one monitor
to another, then its corresponding TaskItem is removed from its
TaskList and added to the TaskList of the other monitor.