windowpicker: cleanup window initialization
This cleans up the initialization for WnckWindows in the
TaskList. There are several cases when we need to initialize
the windows:

 a) When the panel starts and initializes the WindowPicker applets.
 b) When the panel is running but a new WindowPicker applet is
    added to a panel.
 c) When the screen geometry or monitor configuration changes because
    it might cause windows to move to a different monitor.

We are using an idle function to ensure that GTK has finished
initializing the TaskList widget before we initialize the windows
on the list. This is necessary so we can reliably determine whether
a WnckWindow is on the same monitor as the TaskList to which we want
to add the window.

This change also moves the applet setup functions from wp_applet_init
to wp_applet_load to ensure that the TaskList widget has been correctly
initialized when our init_window idle function runs.