window-picker: fix size-allocation issues
Widgets should not call gtk_widget_set_size_request in the size-allocate
callback. This commit adds height-for-width/width-for-height management
to the TaskItem to correctly set the required size of the TaskItem.

We also add handling of the "placement-changed" signal in the TaskItem
to handle changes in the applet orientation.

To avoid clipping of the Tasklist we also update the size_hints of the
applet. Without this TaskItems will have the correct size, but TaskList
will be clipped.

Co-authored-by: Alberts Muktupāvels <alberts.muktupavels@gmail.com>