window-picker: handle item lifecycle in TaskList
When moving items between monitors we were seeing use-after-free errors
because the TaskItem was accessing its WnckWindow after it was already
destroyed. The actual cause however was, that TaskItems were not
correctly removed from a TaskList.

The actual cause for this problem was, that the user-data in the
'task-item-closed' signal handler was not correctly updated when moving
a TaskItem from one list to another. As a result we were removing
TaskItems from the wrong list when the window was closed. This only
happend when the Window was moved to another monitor before it was
closed.

To solve this issue we simplify the TaskItem lifecycle by moving it to
the TaskList. We now use a GHashTable in each TaskList to map open
windows to their respective TaskItems.