window-picker: remove deprecated gdk_cairo_create
We can actually directly draw to the passed cairo context and do not
need to create a new cario context.

The cairo pointer that we created via gdk_cairo_create was behaving
differently from the cairo pointer that is passed to the draw vfunc.

With the cairo context created by gdk_cairo_create, we had to draw at
coordinates relative to the parent widget. With the cairo context
passed to the draw function we can draw relative to the location of the
TaskItem, and thus do not need to specify x and y offsets from the area
in the different cairo operators.