gweather: Set providers list
If all providers are enabled, GWeather tries to use the US weather office (NWS) provider first, which is available only for cities in the US. After that it falls back to METAR, which does not provide forecast information.

Enable only MET_NO and METAR, so the applet will display forecast information from MET_NO, and will use METAR as a fallback.