command: Don't apply the command immediately
If we bind the setting of the command entry, it will be applied immediately while the user is typing the command, and that partial command will run continuously based on the configured interval. This is not so good, so add a new button to apply the change of the command.